requirements-linux.txt
absl-py==2.1.0
immutabledict==4.2.1
openpyxl==3.1.3
PyYAML==6.0.1
pyinstaller==6.6.0
